+++ src/core/account.ts
@@ -6,12 +6,12 @@ import { bytesToHex } from "@noble/hashes/utils";
 import { AccountAddress } from "./account_address";
 import { Hex } from "./hex";
 import { HexInput, SigningScheme } from "../types";
-import { PrivateKey, PublicKey, Signature } from "../crypto/asymmetric_crypto";
+import { PrivateKey, PublicKey, Signature } from "./crypto/asymmetric_crypto";
 import { derivePath } from "../utils/hdKey";
-import { AuthenticationKey } from "../crypto/authentication_key";
-import { Ed25519PrivateKey, Ed25519PublicKey } from "../crypto/ed25519";
-import { Secp256k1PrivateKey, Secp256k1PublicKey } from "../crypto/secp256k1";
-import { MultiEd25519PublicKey } from "../crypto/multi_ed25519";
+import { AuthenticationKey } from "./authentication_key";
+import { Ed25519PrivateKey, Ed25519PublicKey } from "./crypto/ed25519";
+import { Secp256k1PrivateKey, Secp256k1PublicKey } from "./crypto/secp256k1";
+import { MultiEd25519PublicKey } from "./crypto/multi_ed25519";
 
 /**
  * Class for creating and managing account on Aptos network
