+++ src/core/crypto/multi_ed25519.ts
@@ -1,12 +1,12 @@
 // Copyright © Aptos Foundation
 // SPDX-License-Identifier: Apache-2.0
 
-import { Deserializer } from "../bcs/deserializer";
-import { Serializer } from "../bcs/serializer";
+import { Deserializer } from "../../bcs/deserializer";
+import { Serializer } from "../../bcs/serializer";
 import { Ed25519PublicKey, Ed25519Signature } from "./ed25519";
 import { PublicKey, Signature } from "./asymmetric_crypto";
-import { HexInput } from "../types";
-import { Hex } from "../core/hex";
+import { HexInput } from "../../types";
+import { Hex } from "../hex";
 
 /**
  * Represents the public key of a K-of-N Ed25519 multisig transaction.
