+++ src/core/index.ts
@@ -4,3 +4,6 @@
 export * from "./account_address";
 export * from "./common";
 export * from "./hex";
+export * from "./account";
+export * from "./authentication_key";
+export * from "./crypto";
