+++ src/index.ts
@@ -1,6 +1,7 @@
 // Copyright © Aptos Foundation
 // SPDX-License-Identifier: Apache-2.0
 
+export * from "./core";
 export * from "./api";
 export * from "./client";
 export * from "./utils/apiEndpoints";
