+++ src/transactions/authenticator/transaction.ts
@@ -5,9 +5,9 @@
 
 import { Deserializer, Serializer } from "../../bcs";
 import { AccountAddress } from "../../core";
-import { Ed25519PublicKey, Ed25519Signature } from "../../crypto/ed25519";
-import { MultiEd25519PublicKey, MultiEd25519Signature } from "../../crypto/multi_ed25519";
-import { Secp256k1PublicKey, Secp256k1Signature } from "../../crypto/secp256k1";
+import { Ed25519PublicKey, Ed25519Signature } from "../../core/crypto/ed25519";
+import { MultiEd25519PublicKey, MultiEd25519Signature } from "../../core/crypto/multi_ed25519";
+import { Secp256k1PublicKey, Secp256k1Signature } from "../../core/crypto/secp256k1";
 import { TransactionAuthenticatorVariant } from "../../types";
 import { AccountAuthenticator } from "./account";
