+++ src/transactions/transaction_builder/transaction_builder.ts
@@ -12,8 +12,8 @@ import { AptosConfig } from "../../api/aptos_config";
 import { Deserializer } from "../../bcs";
 import { AccountAddress } from "../../core";
 import { Account } from "../../core/account";
-import { Ed25519PublicKey, Ed25519Signature } from "../../crypto/ed25519";
-import { Secp256k1PublicKey, Secp256k1Signature } from "../../crypto/secp256k1";
+import { Ed25519PublicKey, Ed25519Signature } from "../../core/crypto/ed25519";
+import { Secp256k1PublicKey, Secp256k1Signature } from "../../core/crypto/secp256k1";
 import { getInfo } from "../../internal/account";
 import { getLedgerInfo } from "../../internal/general";
 import { getGasPriceEstimation } from "../../internal/transaction";
