+++ tests/unit/account.test.ts
@@ -1,10 +1,17 @@
 // Copyright © Aptos Foundation
 // SPDX-License-Identifier: Apache-2.0
 
-import { Account } from "../../src/core/account";
-import { AccountAddress, Hex } from "../../src/core";
-import { Ed25519PrivateKey, Ed25519PublicKey, Ed25519Signature } from "../../src/crypto/ed25519";
-import { Secp256k1PrivateKey, Secp256k1PublicKey, Secp256k1Signature } from "../../src/crypto/secp256k1";
+import {
+  Account,
+  AccountAddress,
+  Ed25519PrivateKey,
+  Ed25519PublicKey,
+  Ed25519Signature,
+  Hex,
+  Secp256k1PrivateKey,
+  Secp256k1PublicKey,
+  Secp256k1Signature,
+} from "../../src";
 import { SigningScheme } from "../../src/types";
 import { ed25519, secp256k1TestObject, wallet } from "./helper";
