+++ tests/unit/deserializer.test.ts
@@ -1,8 +1,7 @@
 // Copyright © Aptos Foundation
 // SPDX-License-Identifier: Apache-2.0
 
-import { Serializable, Serializer, Deserializer } from "../../src";
-import { AccountAddress } from "../../src/core";
+import { AccountAddress, Serializable, Serializer, Deserializer } from "../../src";
 
 describe("BCS Deserializer", () => {
   it("deserializes a non-empty string", () => {
