+++ tests/unit/multi_ed25519.test.ts
@@ -1,10 +1,15 @@
 // Copyright © Aptos Foundation
 // SPDX-License-Identifier: Apache-2.0
 
-import { Deserializer, Serializer } from "../../src";
-import { Hex } from "../../src/core";
-import { Ed25519PublicKey, Ed25519Signature } from "../../src/crypto/ed25519";
-import { MultiEd25519PublicKey, MultiEd25519Signature } from "../../src/crypto/multi_ed25519";
+import {
+  Deserializer,
+  Ed25519PublicKey,
+  Ed25519Signature,
+  Hex,
+  MultiEd25519PublicKey,
+  MultiEd25519Signature,
+  Serializer,
+} from "../../src";
 import { multiEd25519PkTestObject, multiEd25519SigTestObject } from "./helper";
 
 describe("MultiPublicKey", () => {
