+++ tests/unit/secp256k1.test.ts
@@ -1,11 +1,9 @@
 // Copyright © Aptos Foundation
 // SPDX-License-Identifier: Apache-2.0
 
-import { Hex } from "../../src/core";
-import { Secp256k1PrivateKey, Secp256k1PublicKey, Secp256k1Signature } from "../../src/crypto/secp256k1";
 import { secp256k1 } from "@noble/curves/secp256k1";
 import { secp256k1TestObject } from "./helper";
-import { Deserializer, Serializer } from "../../src";
+import { Deserializer, Hex, Secp256k1PrivateKey, Secp256k1PublicKey, Secp256k1Signature, Serializer } from "../../src";
 
 describe("Secp256k1PublicKey", () => {
   it("should create the instance correctly without error", () => {
