+++ tests/unit/type_tag.test.ts
@@ -18,8 +18,7 @@ import {
   TypeTagU8,
   TypeTagVector,
 } from "../../src/transactions/typeTag/typeTag";
-import { Deserializer, Serializer } from "../../src";
-import { AccountAddress, AddressInvalidReason, ParsingError } from "../../src/core";
+import { AccountAddress, AddressInvalidReason, Deserializer, ParsingError, Serializer } from "../../src";
 
 const expectedTypeTag = {
   string: "0x1::some_module::SomeResource",
